+++ test/index.test.js
@@ -1,5 +1,0 @@
-// TODO: write tests
-
-it('test', () => {
-  expect(1).toEqual(1)
-})